15 missing packages in codespaces instance (#16)
* Added a devcontainer.json file to configure codespaces instance.
Added cspell extension files to .gitignore

* Moved devcointer.json to .devcontainer folder.

* Run apt-get as sudo
Rename development container

* Now store multiple data_frames in a data object, as there can be more than one table.
Each data-frame is stored in a dict with name of the table in the database as the key.
No longer auto-generating the table names. It's now bespoke for each data file.

* Moved the class for processing and storing title.basics.tsv to a new file.
Finally fixed the issue with not filtering adult titles - seems casting to integer doesn't work very well.

* Removed python script for downloading data from IMDb.

* Updated CLI instructions for load_imdb.py

* Ignore uncompressed sqlite database.

* Installing packages using devcontainer.json features instead of apt-get.
Installing git-lfs as an apt package, because devcontainer feature was throwing errors.
Uncompressing the sqlite database.

* Added sqlite3 package to apt-packages feature in devcontainer.json

* Restored accidentally removed compressed file.
Keep the compressed DB after decompressing it

* Adding workspace settings.json as it has details of the SQLTools connection to the imdb.sqlite database.

* Fixed SQLTools connection